requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backports.zoneinfo==0.2.1
billiard==4.2.0
bitarray==2.8.2
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
cytoolz==0.12.2
Django==4.2.7
django-compat==1.0.15
django-environ==0.11.2
django-timezone-field==6.0.1
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.2
eth-utils==2.3.1
frozenlist==1.4.0
hexbytes==0.3.1
idna==3.4
importlib-resources==6.1.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kombu==5.3.3
lru-dict==1.2.0
multidict==6.0.4
parsimonious==0.9.0
pkgutil-resolve-name==1.3.10
prompt-toolkit==3.0.39
protobuf==4.25.0
psycopg2==2.9.9
pycryptodome==3.19.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pyunormalize==15.0.0
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rlp==3.0.0
rpds-py==0.12.0
six==1.16.0
sqlparse==0.4.4
toolz==0.12.0
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.9
web3==6.11.3
websockets==12.0
yarl==1.9.2
zipp==3.17.0
